% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.beta.MH.RW.glm}
\alias{.beta.MH.RW.glm}
\title{Beta MH RW sampler from freq PEM fit}
\usage{
.beta.MH.RW.glm(df, beta, beta_count, cprop_beta)
}
\arguments{
\item{df}{Data frame with indicators}

\item{beta}{vector of parameters}

\item{beta_count}{count number of accepted proposals}

\item{cprop_beta}{proposal scalar}
}
\value{
beta, either old or new move
}
\description{
Sample beta from RW sampler
}
