% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.beta_MH_MALA}
\alias{.beta_MH_MALA}
\title{Proposal beta with a Metrolopis Adjusted Langevin (MALA)}
\usage{
.beta_MH_MALA(df, beta, bp, cprop_beta, count_beta)
}
\arguments{
\item{df}{Data frame with indicators}

\item{beta}{vector of parameters}

\item{bp}{number of covariates}

\item{cprop_beta}{proposal variance standard deviation}

\item{count_beta}{count number of accepts}
}
\value{
updated beta vector
}
\description{
Proposal beta with a Metrolopis Adjusted Langevin (MALA)
}
