% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.beta_MH_NR}
\alias{.beta_MH_NR}
\title{Newton Raphson MH move}
\usage{
.beta_MH_NR(df, beta, bp, cprop_beta, beta_count)
}
\arguments{
\item{df}{Data frame with indicators}

\item{beta}{vector of parameters}

\item{bp}{number of covariates}

\item{cprop_beta}{proposal scalar}

\item{beta_count}{count number of accepts}
}
\value{
updated beta
}
\description{
Sample beta from RW sampler
}
