% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.beta_MH_RW}
\alias{.beta_MH_RW}
\title{Beta Metropolis-Hastings Random walk move}
\usage{
.beta_MH_RW(df, beta, bp, cprop_beta, beta_count)
}
\arguments{
\item{df}{data.frame from dataframe_fun()}

\item{beta}{beta values}

\item{bp}{number of covariates}

\item{cprop_beta}{hyperparameter for beta proposal standard deviation}

\item{beta_count}{number of moves done for beta}
}
\value{
beta, either old or new move
}
\description{
Update beta via a Metropolis-Hastings Random Walk move
}
