% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.beta_mom.NR.fun}
\alias{.beta_mom.NR.fun}
\title{First and second derivative of target for mode and variance of proposal}
\usage{
.beta_mom.NR.fun(df, k, beta, bp, cprop_beta)
}
\arguments{
\item{df}{Data frame with indicators}

\item{k}{index}

\item{beta}{vector of parameters}

\item{bp}{number of covariates}

\item{cprop_beta}{proposal variance standard deviation}
}
\value{
First and second derivative mode and variance
}
\description{
First and second derivative of target for mode and variance of proposal
}
