% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.beta_mom}
\alias{.beta_mom}
\title{Mean for MALA using derivative for beta proposal}
\usage{
.beta_mom(df, k, beta, bp, cprop_beta)
}
\arguments{
\item{df}{Data frame with indicators}

\item{k}{index for beta}

\item{beta}{vector of parameters}

\item{bp}{number of covariates}

\item{cprop_beta}{proposal standard dev}
}
\value{
proposal mean
}
\description{
Mean for MALA using derivative for beta proposal
}
