% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJMCMC.R
\name{.birth_move}
\alias{.birth_move}
\title{Birth move in RJMCMC}
\usage{
.birth_move(U, sj, s_star, sjm1, x, j)
}
\arguments{
\item{U}{uniform random number}

\item{sj}{upcoming split point location, j}

\item{s_star}{new split point location, *}

\item{sjm1}{previous split point location, j-1}

\item{x}{vector of parameter values, length J + 1}

\item{j}{split point}
}
\value{
vector with adjusted parameter values after additional split point,
length J + 2
}
\description{
Calculates new values of x when proposing another split point, based on a 
weighted mean, as x_new/x <- (1-U)/U
}
