% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dataframe_fun}
\alias{.dataframe_fun}
\title{Create data.frame for piecewise exponential models}
\usage{
.dataframe_fun(Y, I, X, s, lambda, bp, J)
}
\arguments{
\item{Y}{time-to-event}

\item{I}{censor indicator}

\item{X}{design Matrix}

\item{s}{split point locations, including start and end (length J+2)}

\item{lambda}{baseline Hazards (length J+1)}

\item{bp}{number of covariates}

\item{J}{number of split points}
}
\value{
data.frame with columns c(tstart, id, X1,..., Xp, Y, I, lambda)
}
\description{
Construct a split data.frame for updated split points
}
