% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJMCMC.R
\name{.death_move}
\alias{.death_move}
\title{Death move in RJMCMC}
\usage{
.death_move(sjp1, sj, sjm1, x, j)
}
\arguments{
\item{sjp1}{upcoming split point location, J + 1}

\item{sj}{split point location to be removed, j}

\item{sjm1}{previous split point location, j-1}

\item{x}{vector of parameter values, length J + 1}

\item{j}{split point}
}
\value{
vector with adjusted parameter values after removal of split point,
length J
}
\description{
Calculates new values of x when proposing the death of a split point
}
