% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.glmFit}
\alias{.glmFit}
\title{Fit frequentist piecewise exponential model for MLE and information matrix of beta}
\usage{
.glmFit(df)
}
\arguments{
\item{df}{Data frame with time-to-event, censoring indicator and covariates}
}
\value{
beta MLE and inverse of information matrix
}
\description{
Compute MLE for PEM
}
