% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_updates.R
\name{.ICAR_calc}
\alias{.ICAR_calc}
\title{Calculate covariance matrix in the MVN-ICAR}
\usage{
.ICAR_calc(s, J, clam)
}
\arguments{
\item{s}{split points, J + 2}

\item{J}{number of split points}

\item{clam}{controls neighbor interactions, in range (0, 1)}
}
\value{
Sigma_s = (I - W)^(-1) * Q, W, Q
}
\description{
Calculate covariance matrix in the MVN-ICAR
}
