% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.input_check}
\alias{.input_check}
\title{Input checker}
\usage{
.input_check(
  Y,
  Y_0,
  X,
  X_0,
  tuning_parameters,
  initial_values = NULL,
  hyperparameters
)
}
\arguments{
\item{Y}{current time-to-event data}

\item{Y_0}{historical time-to-event data}

\item{X}{design Matrix}

\item{X_0}{design Matrix for historical data}

\item{tuning_parameters}{list of tuning parameters}

\item{initial_values}{list of initial values (optional)}

\item{hyperparameters}{list of hyperparameters}
}
\value{
a print statement
}
\description{
Checks inputs before Gibbs sampler is run
}
