% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_updates.R
\name{.lambda_conj_prop}
\alias{.lambda_conj_prop}
\title{Propose lambda from a gamma conditional conjugate posterior proposal}
\usage{
.lambda_conj_prop(df, beta, j, bp, alam = 0.01, blam = 0.01)
}
\arguments{
\item{df}{data.frame from dataframe_fun()}

\item{beta}{parameter value for beta}

\item{j}{current split point}

\item{bp}{number of covariates}

\item{alam}{lambda hyperparameter, default set to 0.01}

\item{blam}{lambda hyperparameter, default set to 0.01}
}
\value{
list containing proposed lambda, shape and rate parameters
}
\description{
Propose lambda from a gamma conditional conjugate posterior proposal
}
