% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_updates.R
\name{.lgamma_ratio}
\alias{.lgamma_ratio}
\title{Calculate log gamma ratio for two different parameter values}
\usage{
.lgamma_ratio(x1, x2, shape, rate)
}
\arguments{
\item{x1}{old parameter value}

\item{x2}{proposed parameter value}

\item{shape}{shape parameter}

\item{rate}{rate parameter}
}
\value{
log gamma ratio
}
\description{
Calculate log gamma ratio for two different parameter values
}
