% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.llikelihood_ratio_beta}
\alias{.llikelihood_ratio_beta}
\title{Loglikelihood ratio calculation for beta parameters}
\usage{
.llikelihood_ratio_beta(df, beta, beta_new)
}
\arguments{
\item{df}{data.frame from dataframe_fun()}

\item{beta}{beta values}

\item{beta_new}{proposed beta values}
}
\value{
likelihood ratio
}
\description{
Compute log likelihood for beta update
}
