% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_updates.R
\name{.llikelihood_ratio_lambda}
\alias{.llikelihood_ratio_lambda}
\title{Log likelihood for lambda / lambda_0 update}
\usage{
.llikelihood_ratio_lambda(df, df_prop, beta)
}
\arguments{
\item{df}{data.frame from dataframe_fun()}

\item{df_prop}{proposal data.frame}

\item{beta}{parameter value for beta}
}
\value{
log likelihood ratio for lambda
}
\description{
Log likelihood for lambda / lambda_0 update
}
