% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.log_likelihood}
\alias{.log_likelihood}
\title{Log likelihood function}
\usage{
.log_likelihood(df, beta)
}
\arguments{
\item{df}{data.frame containing data, time split points, and lambda}

\item{beta}{coefficients for covariates}
}
\value{
log likelihood given lambdas and betas
}
\description{
Log likelihood function
}
