% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.logsumexp}
\alias{.logsumexp}
\title{Computes the logarithmic sum of an exponential}
\usage{
.logsumexp(x)
}
\arguments{
\item{x}{set of log probabilities}
}
\value{
the logarithmic sum of an exponential
}
\description{
Computes the logarithmic sum of an exponential
}
