% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.lprop.dens.beta.NR}
\alias{.lprop.dens.beta.NR}
\title{log Gaussian proposal density for Newton Raphson proposal}
\usage{
.lprop.dens.beta.NR(beta.prop, mu_old, var_old)
}
\arguments{
\item{beta.prop}{beta proposal}

\item{mu_old}{density mean}

\item{var_old}{density variance}
}
\value{
log Gaussian density
}
\description{
log Gaussian proposal density for Newton Raphson proposal
}
