% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_updates.R
\name{.lprop_density_beta}
\alias{.lprop_density_beta}
\title{Log density of proposal for MALA}
\usage{
.lprop_density_beta(beta_prop, mu, cprop_beta)
}
\arguments{
\item{beta_prop}{proposal beta}

\item{mu}{mean of proposal distribution}

\item{cprop_beta}{proposal standard dev}
}
\value{
log density
}
\description{
Log density of proposal for MALA
}
