% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_updates.R
\name{.mu_update}
\alias{.mu_update}
\title{Calculate mu posterior update}
\usage{
.mu_update(Sigma_s, lambda_0, sigma2, J)
}
\arguments{
\item{Sigma_s}{VCV matrix (j + 1) x (j + 1).}

\item{lambda_0}{Baseline hazard.}

\item{sigma2}{Scale variance.}

\item{J}{Number of split point.}
}
\value{
mu update from Normal.
}
\description{
Calculate mu posterior update
}
