% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.normalize_prob}
\alias{.normalize_prob}
\title{Normalize a set of probability to one, using the the log-sum-exp trick}
\usage{
.normalize_prob(x)
}
\arguments{
\item{x}{set of log probabilities}
}
\value{
normalized set of log probabilities
}
\description{
Normalize a set of probability to one, using the the log-sum-exp trick
}
