% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_updates.R
\name{.nu_sigma_update}
\alias{.nu_sigma_update}
\title{Calculates nu and sigma2 for the Gaussian Markov random field prior, 
for a given split point j}
\usage{
.nu_sigma_update(j, lambda_0, mu, sigma2, W, Q, J)
}
\arguments{
\item{j}{current split point}

\item{lambda_0}{historical baseline hazard}

\item{mu}{prior mean for baseline hazard}

\item{sigma2}{prior variance hyperparameter for baseline hazard}

\item{W}{influence from right and left neighbors}

\item{Q}{individual effect of neighborhood}

\item{J}{number of split points}
}
\value{
nu and sigma2
}
\description{
Calculates nu and sigma2 for the Gaussian Markov random field prior, 
for a given split point j
}
