% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.plot_hist}
\alias{.plot_hist}
\title{Plot histogram from MCMC samples}
\usage{
.plot_hist(
  samples,
  title = "",
  xlab = "Values",
  ylab = "Frequency",
  color = "black",
  fill = "blue",
  binwidth = 0.05,
  scale_x = FALSE
)
}
\arguments{
\item{samples}{data.frame containing the discrete MCMC samples}

\item{title}{title of the plot, default is none}

\item{xlab}{x-label of the plot, default is "Values"}

\item{ylab}{y-label of the plot, default is "Frequency"}

\item{color}{outline color for the bars, default is "black"}

\item{fill}{fill color, default is "blue"}

\item{binwidth}{width of the histogram bins, default is 0.5}

\item{scale_x}{option to scale the x-axis, suitable for discrete samples, 
default is FALSE}
}
\value{
a ggplot2 object
}
\description{
Plots a histogram of the given discrete MCMC samples
}
