% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.plot_trace}
\alias{.plot_trace}
\title{Plot MCMC trace}
\usage{
.plot_trace(
  x_lim,
  samples,
  title = "",
  xlab = "",
  ylab = "",
  color = "black",
  linewidth = 1
)
}
\arguments{
\item{x_lim}{x-axis of the plot}

\item{samples}{samples from MCMC}

\item{title}{optional, add title to plot}

\item{xlab}{optional, add xlabel}

\item{ylab}{optional, add ylabel}

\item{color}{color of the mid line, default is black}

\item{linewidth}{thickness of the plotted line, default is 1}
}
\value{
a ggplot2 object
}
\description{
Creates a trace plot of given MCMC samples.
}
