% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.predictive_hazard}
\alias{.predictive_hazard}
\title{Predictive hazard from BayesFBHborrow object}
\usage{
.predictive_hazard(out_slam, x_pred, beta_samples)
}
\arguments{
\item{out_slam}{samples from the smoothed baseline hazard}

\item{x_pred}{set of predictors to be used for calculating the predictive hazard}

\item{beta_samples}{samples of the covariates}
}
\value{
matrix of the predictive hazard
}
\description{
Predictive hazard from BayesFBHborrow object
}
