% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.predictive_hazard_ratio}
\alias{.predictive_hazard_ratio}
\title{Predictive hazard ratio (HR) from BayesFBHborrow object}
\usage{
.predictive_hazard_ratio(x_pred, beta_samples)
}
\arguments{
\item{x_pred}{set of predictors to be used for calculating the predictive HR}

\item{beta_samples}{samples of the covariates}
}
\value{
posterior samples for expectation and credible intervals
}
\description{
Predictive hazard ratio (HR) from BayesFBHborrow object
}
