% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.predictive_survival}
\alias{.predictive_survival}
\title{Predictive survival from BayesFBHborrow object}
\usage{
.predictive_survival(grid_width, out_slam, x_pred, beta_samples)
}
\arguments{
\item{grid_width}{size of time step}

\item{out_slam}{samples from the smoothed baseline hazard}

\item{x_pred}{set of predictors to be used for calculating the predictive survival}

\item{beta_samples}{samples of the covariates}
}
\value{
matrix of the predictive survival
}
\description{
Predictive survival from BayesFBHborrow object
}
