% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.set_hyperparameters}
\alias{.set_hyperparameters}
\title{Set tuning parameters}
\usage{
.set_hyperparameters(hyperparameters = NULL, model_choice)
}
\arguments{
\item{hyperparameters}{list of hyperparameters, could contain any combination
of the listed hyperparameters}

\item{model_choice}{choice of model, could be either of 'mix', 'uni' or 'all'}
}
\value{
filled list of tuning_parameters
}
\description{
Set tuning parameters
}
