% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_updates.R
\name{.shuffle_split_point_location_NoBorrow}
\alias{.shuffle_split_point_location_NoBorrow}
\title{Metropolis Hastings step: shuffle the split point locations (without 
Bayesian borrowing)}
\usage{
.shuffle_split_point_location_NoBorrow(
  df,
  Y_0,
  I_0,
  X_0,
  lambda_0,
  beta_0,
  s,
  J,
  bp_0,
  clam_smooth
)
}
\arguments{
\item{df}{dataframe containing trial data and parameters}

\item{Y_0}{data}

\item{I_0}{censoring indicator}

\item{X_0}{design matrix}

\item{lambda_0}{baseline hazard}

\item{beta_0}{parameter vector}

\item{s}{split point locations, J + 2}

\item{J}{number of split points}

\item{bp_0}{number of covariates in historical trial}

\item{clam_smooth}{neighbor interactions, in range (0, 1), for ICAR update}
}
\value{
list containing new split points, updated Sigma_s and data.frames
for historic and current trial data
}
\description{
Metropolis Hastings step: shuffle the split point locations (without 
Bayesian borrowing)
}
