% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.smooth_hazard}
\alias{.smooth_hazard}
\title{Smoothed hazard function}
\usage{
.smooth_hazard(out_slam, beta_samples = NULL)
}
\arguments{
\item{out_slam}{samples from GibbsMH of the baseline hazard}

\item{beta_samples}{samples from GibbsMH from the treatment effect}
}
\value{
smoothed function for the baseline hazard
}
\description{
Smoothed hazard function
}
