% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.smooth_survival}
\alias{.smooth_survival}
\title{Smoothed survival curve}
\usage{
.smooth_survival(grid_width, out_slam, beta_samples = NULL)
}
\arguments{
\item{grid_width}{step size}

\item{out_slam}{samples from GibbsMH of the baseline hazard}

\item{beta_samples}{samples from GibbsMH from the treatment effect}
}
\value{
smoothed survival function
}
\description{
Smoothed survival curve
}
