% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_updates.R
\name{.tau_update}
\alias{.tau_update}
\title{Sample tau from posterior distribution}
\usage{
.tau_update(
  lambda_0,
  lambda,
  J,
  s,
  a_tau,
  b_tau,
  c_tau = NULL,
  d_tau = NULL,
  p_0 = NULL,
  type
)
}
\arguments{
\item{lambda_0}{historical baseline hazard}

\item{lambda}{baseline hazard}

\item{J}{number of split points}

\item{s}{split point locations, J + 2}

\item{a_tau}{Inverse Gamma hyperparameter}

\item{b_tau}{Inverse Gamma hyperparameter}

\item{c_tau}{Inverse Gamma hyperparameter}

\item{d_tau}{Inverse Gamma hyperparameter}

\item{p_0}{mixture ratio}

\item{type}{choice of borrowing, "mix", "uni", or any other string for 
borrowing on every baseline hazard without mixture}
}
\value{
list containing tau and new mixture ratio
}
\description{
Sample tau from posterior distribution
}
