% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsMH.R
\name{GibbsMH}
\alias{GibbsMH}
\title{S3 generic, calls the correct GibbsMH sampler}
\usage{
GibbsMH(
  Y,
  I,
  X,
  Y_0 = NULL,
  I_0 = NULL,
  X_0 = NULL,
  tuning_parameters,
  initial_values,
  hyperparameters,
  lambda_hyperparameters,
  iter,
  warmup_iter,
  refresh,
  max_grid
)
}
\arguments{
\item{Y}{data}

\item{I}{event indicator}

\item{X}{design matrix}

\item{Y_0}{historical data, default is NULL}

\item{I_0}{historical event indicator, default is NULL}

\item{X_0}{historical design matrix, default is NULL}

\item{tuning_parameters}{list of "cprop_beta", "cprop_beta_0", "alpha", "Jmax",
and "pi_b"}

\item{initial_values}{list containing the initial values of c("J", "s_r",
"mu", "sigma2", "tau", "lambda_0", "lambda", "beta_0", "beta") (optional)}

\item{hyperparameters}{list containing the hyperparameters c("a_tau", "b_tau",
 "c_tau", "d_tau","type", "p_0", "a_sigma", "b_sigma", "Jmax", "clam_smooth",
 "cprop_beta", "phi", "pi_b"). Default is list("a_tau" = 1,"b_tau" = 1,"c_tau" = 1,
"d_tau" = 0.001, "type" = "mix", "p_0" = 0.5, "a_sigma" = 2, "b_sigma" = 2,
"Jmax" = 20, "clam_smooth" = 0.8, "cprop_beta" = 0.3, "phi" = 3, "pi_b" = 0.5)}

\item{lambda_hyperparameters}{contains two hyperparameters (a_lambda and b_lambda) used 
for the update of lambda and lambda_0}

\item{iter}{number of iterations for MCMC sampler, excluding warmup, 
default is 2000}

\item{warmup_iter}{number of warmup iterations (burn-in) for MCMC sampler, 
default is 2000}

\item{refresh}{number of iterations between printed screen updates, 
default is 500}

\item{max_grid}{grid size for the smoothed baseline hazard, default is 2000}
}
\value{
depending on if the user wishes to borrow; returns a list with values
 after each iteration for parameters: out_fixed (J, mu, sigma2, beta), lambda,
 lambda_0, tau, s, as well as tuning values of the total number of accepts:
 lambda_move, lambda_0_move and beta_move. Also included is the out_slam which
 contains the shrunk estimate of the baseline hazard.
}
\description{
An MCMC sampler for Bayesian borrowing with time-to-event data. 
We obtain a flexible baseline hazard function by making the split points 
random within a piecewise exponential model and using a Gaussian Markov 
random field prior to smooth the baseline hazards. Only calls the sampler and
does not run any input checks. Best practice is to call BayesFBHborrow(), if the 
user is not familiar with the model at hand.
}
\examples{
set.seed(123)
# Load example data and set your initial values and hyper parameters
data(weibull_cc, package = "BayesFBHborrow")
data(weibull_hist, package = "BayesFBHborrow")

# The datasets consists of 3 (2) columns named "tte", "event" and "X" 
# (only for concurrent). To explicitly run the sampler, extract the samples as
# following
Y <- weibull_cc$tte
I <- weibull_cc$event
X <- matrix(weibull_cc$X_trt)

Y_0 <- weibull_hist$tte
I_0 <- weibull_hist$event
X_0 <- NULL

# Specify hyperparameters and tuning parameters
hyper <-  list("a_tau" = 1, 
               "b_tau" = 0.001,
               "c_tau" = 1,
               "d_tau" = 1, 
               "type" = "all",
               "p_0" = 0.5, 
               "a_sigma" = 2,
               "b_sigma" = 2,
               "clam_smooth" = 0.5,
               "phi" = 3)

tuning_parameters <- list("Jmax" = 5,
                          "pi_b" = 0.5,
                          "cprop_beta" = 0.5,
                          "alpha" = 0.4)
                          
# Set initial values to 'NULL' for default settings
output <- GibbsMH(Y, I, X, Y_0, I_0, X_0,
                  tuning_parameters, initial_values = NULL, hyper, 
                  iter = 5, warmup_iter = 1)
                  
}
