\docType{import}
\name{reexports}
\alias{reexports}
\alias{Surv}
\alias{ci}
\alias{kable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{survival}{\code{\link[survival]{Surv}}}
  \item{bayestestR}{\code{\link[bayestestR]{ci}}}
  \item{kableExtra}{\code{\link[knitr]{kable}}}
}}

