% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.input_check}
\alias{.input_check}
\title{Input checker}
\usage{
.input_check(Y, Y_0, X, X_0, hyperparameters, initial_param)
}
\arguments{
\item{Y}{current time-to-event data}

\item{Y_0}{historical time-to-event data}

\item{X}{design Matrix}

\item{X_0}{design Matrix for historical data}

\item{hyperparameters}{list of hyperparameters}

\item{initial_param}{list of initial parameters}
}
\value{
a print statement
}
\description{
Checks inputs before Gibbs sampler is run
}
